+++ bip/base/__init__.py
@@ -1,5 +1,6 @@
-from .utils import get_highlighted_identifier_as_int, Ptr, get_ptr_size, relea, absea, get_addr_by_name, get_name_by_addr, get_struct_from_lvar, bip_exec_sync, min_ea, max_ea
-from .bipelt import BipBaseElt, BipRefElt, BipElt, GetElt, GetEltByName, Here
+from .bipidb import BipIdb, min_ea, max_ea, Here
+from .bipida import BipIda
+from .bipelt import BipBaseElt, BipRefElt, BipElt, GetElt, GetEltByName
 from .instr import BipInstr
 from .data import BipData
 from .operand import BipOpType, BipDestOpType, BipOperand
@@ -10,3 +11,4 @@ from .biperror import BipError, BipDecompileError
 from .func import _BipFuncFlags, BipFunction
 from .block import BipBlockType, BipBlock
 from .biptype import BipType, BTypeEmpty, BTypePartial, BTypeVoid, BTypeInt, BTypeBool, BTypeFloat, BTypePtr, BTypeArray, BTypeFunc, BTypeStruct, BTypeUnion, BTypeEnum
+