+++ bip/gui/__init__.py
@@ -2,5 +2,6 @@ from .old_actions import *
 from .activity import BipActivity, BipActivityContainer
 from .actions import BipAction
 from .menutb import *
+from .userselect import BipUserSelect
 from .pluginmanager import get_plugin_manager, BipPluginLoader
 from .plugin import BipPlugin, shortcut, menu