+++ doc/general/changelog.rst
@@ -8,6 +8,14 @@ New features may be listed but probably not always.
 Change from v0.3 to v1.0
 ========================
 
+New features:
+
+* new class :class:`bip.base.BipIdb`
+* new class :class:`bip.base.BipIda`
+* new class :class:`bip.gui.BipUserSelect`
+
+Breaking changes:
+
 * renaming of ``bip.base.XrefTypes`` to ``bip.base._XrefTypes``
 * renaming of ``bip.base.DestOpType`` to ``bip.base.BipDestOpType``
 * renaming of ``bip.base.OpType`` to ``bip.base.BipOpType``
@@ -35,7 +43,16 @@ Change from v0.3 to v1.0
 * renaming of ``bip.hexrays.CNode.cfunc`` to ``bip.hexrays.CNode.hxcfunc``
 * renaming of ``bip.base.BipFunction.hxfunc`` to ``bip.base.BipFunction.hxcfunc``
 * renaming of ``bip.hexrays.HxLvar.hxfunc`` to ``bip.hexrays.HxLvar.hxcfunc``
-
+* function ``bip.base.utils.get_ptr_size`` became static method ``bip.base.BipIdb.ptr_size``
+* function ``bip.base.utils.absea`` became a static method of ``bip.base.BipIdb``
+* function ``bip.base.utils.relea`` became a static method of ``bip.base.BipIdb``
+* ``min_ea``, ``max_ea`` and ``Here``, functions are now in ``bip.base.bipidb``
+* function ``bip.base.utils.get_addr_by_name`` has been removed.
+* function ``bip.base.utils.get_name_by_addr`` has been removed.
+* function ``bip.base.utils.get_struct_from_lvar`` has been removed.
+* function ``bip.base.utils.Ptr`` became static method of ``bip.base.BipData.get_ptr``
+* function ``bip.base.utils.bip_exec_sync`` became static method ``bip.base.BipIda.exec_sync``
+* function ``bip.base.utils.get_highlighted_identifier_as_int`` became static method ``BipUserSelect.get_curr_highlighted_int``
 
 Sed script for automatic update of plugins (no garantee to be perfect or to
 avoid colisions) (use with ``sed -f RULEFILE INPUTFILE``):
@@ -62,9 +79,13 @@ avoid colisions) (use with ``sed -f RULEFILE INPUTFILE``):
     s/_createChild/_create_child/g
     s/st_childs/stmt_childs/g
     s/GetCNode/from_citem/g
+    s/get_ptr_size/BipIdb.ptr_size/g
+    s/bip_exec_sync/BipIda.exec_sync/g
+    s/get_highlighted_identifier_as_int/BipUserSelect.get_curr_highlighted_int/g
 
 Are not included in this sed file the change to ``BipInstr.Make``,
-``BipFunction.Count``, ``Cnode.cfunc`` which can easilly create problems.
+``BipFunction.Count``, ``Cnode.cfunc``, ``Ptr`` which can easilly create
+problems.